aggiunta TH alle pagine html

+++ src/main/java/com/example/Pista/dao/PilotaDao.java
@@ -3,5 +3,5 @@ package com.example.Pista.dao;
 import com.example.Pista.model.Pilota;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PilotaDao extends CrudRepository<Pilota, Long> {
+public interface PilotaDao extends CrudRepository<Pilota, Integer> {
 }
